requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
sqlalchemy==2.0.23
python-multipart==0.0.6
bcrypt==4.0.1
pyjwt==2.8.0
cryptography==41.0.8
python-dotenv==1.0.0
email-validator==2.1.0
passlib[bcrypt]==1.7.4
gunicorn==21.2.0